Update parameter type from Long to int in countFlightsByAirport

+++ ProjectBookingFlights/src/main/java/com/airport/ProjectBookingFlights/controllers/AirportController.java
@@ -37,7 +37,7 @@ public class AirportController {
     }
 
     @GetMapping("/{id}/flights/count")
-    public Long countFlightsByAirport(@PathVariable Long id) {
+    public int countFlightsByAirport(@PathVariable int id) {
         return airportService.countFlightsByAirportId(id);
     }
